+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added "Give Yoyo" effect (disabled).
+
 ## [2.0.11] - 2023-12-13
 
 - Changed "Do-or-die" effect to no longer include "Sleep" challenge.